% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{cropresamplenc}
\alias{cropresamplenc}
\title{Resample nc file to spatial array}
\usage{
cropresamplenc(file, r, Trace = TRUE)
}
\arguments{
\item{file}{the name of the .nc file which the data are to be read from.
If it does not contain an absolute path, the file name is relative to the
current working directory, \code{\link[=getwd]{getwd()}}. Tilde-expansion is performed where
supported.}

\item{r}{an object of type raster. The returned \code{spatialarray} has the same
resolution and extent as \code{r}.}

\item{Trace}{optional logical indicating whether to create and plot
a raster of every 100th entry to enable progress to be tracked.}
}
\value{
an object of class \code{spatialarray}. An object of class \code{spatialarray} is
a list containing the following components:
\describe{
\item{arraydata}{A three-dimensional array of values corresponding to values of the resampled .nc file}
\item{times}{An object of class \code{POSIXlt} of times \code{arraydata}}
\item{crs}{An object of class \code{crs} indicating the coordinate reference system associated with \code{arraydata}}
\item{extent}{An object of class \code{extent} indicating the geographic extent covered by \code{arraydata}}
\item{units}{Units of \code{arraydata}}
\item{description}{Description of \code{arraydata}}
}
}
\description{
Resample nc file to spatial array
}
