% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem5k}
\alias{dem5k}
\title{Digital elevation dataset for Cornwall}
\format{A raster with 23 rows and 34 columns:}
\usage{
dem5k
}
\description{
A 5 km resolution raster of elevations in Cornwall
}
\keyword{datasets}
