% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem60k}
\alias{dem60k}
\title{Digital elevation dataset for UK}
\format{A raster with 21 rows and 12 columns:}
\usage{
dem60k
}
\description{
A 60 km resolution raster of elevations in the UK
}
\keyword{datasets}
