% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demdif}
\alias{demdif}
\title{Elevation differences between 60 km and 5km resolution data}
\format{A raster with 23 rows and 34 columns:}
\usage{
demdif
}
\description{
A 5 km resolution raster of elevation differences between 60 km and 5km resolution data
}
\keyword{datasets}
