% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{.rasterTps}
\alias{.rasterTps}
\title{Perform thin-plate spline with rasters}
\usage{
.rasterTps(y, rc1, rf1, rc2 = NA, rf2 = NA, rc3 = NA, rf3 = NA)
}
\description{
Perform thin-plate spline with rasters
}
