% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.resamplearray}
\alias{.resamplearray}
\title{resample spatial array object to resolution of r}
\usage{
.resamplearray(ae, r, tme, Trace)
}
\description{
resample spatial array object to resolution of r
}
