% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.timesel}
\alias{.timesel}
\title{creates sequence matching 10 years of UKCP18 data, with missing days ommited}
\usage{
.timesel(startyear)
}
\description{
creates sequence matching 10 years of UKCP18 data, with missing days ommited
}
