% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.tmecreate}
\alias{.tmecreate}
\title{Create POSIXlt object of 3600 daily times for a  10 year UKCP18 dataset}
\usage{
.tmecreate(startyear)
}
\description{
Create POSIXlt object of 3600 daily times for a  10 year UKCP18 dataset
}
