% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerfunctions.R
\name{gamcorrect}
\alias{gamcorrect}
\title{Fit General Additive Model to correct UKCP18 data}
\usage{
gamcorrect(observed, ukcp, Trace = TRUE, positive = TRUE)
}
\arguments{
\item{observed}{object of class \code{spatialarray} of observed climate data}

\item{ukcp}{object of class \code{spatialarray} of UKCP18 climate data for the
corresponding period and location}

\item{Trace}{optional logical indicating whether to plot model fit}

\item{positive}{optional logical indicating whether all values are positive (as
with e.g. radiation)}
}
\value{
an object of class \code{gam}.
}
\description{
Fit General Additive Model to correct UKCP18 data
}
\details{
This function ranks the values of \code{observed} and \code{ukcp} and fits
a General Additive Model (GAM) to these data. The area and time period covered by
\code{observed} and \code{ukcp}, and the spatial resolution of the data should be
identical. If the datasets contain more than 10,000 values then 10000 values
spanning the full range of data are selected. If \code{positive} is TRUE, then only
values greater than 0 are selected when fitting the GAM.
}
\examples{
# Takes a few seconds to run
modfit <- radfit(sis_2005)
# Takes a few minutes to run
ukcpsishourly <- hourlysis(sis_ukcpdaily, modfit, 2000)
# Select data for 2005
tme <- ukcpsishourly$times
sel <- which(tme$year + 1900 == 2005)
xx <-  ukcpsishourly$arraydata[,,sel]
ukcp_2005 <- list(arraydata = xx, times = tme[sel], crs = ukcpsishourly$crs,
                 extent = ukcpsishourly$extent, units = ukcpsishourly$array.units,
                 description = ukcpsishourly$array.description)
class(ukcp_2005) <- "spatialarray"
od_obs <- opticaldepth(sis_2005)
od_pre <- opticaldepth(ukcp_2005)
gam.mod <- gamcorrect(od_obs, od_pre)
}
