% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{hourlypsl}
\alias{hourlypsl}
\title{Calculate hourly corrected sea-level pressure from daily data}
\usage{
hourlypsl(ukcp_psl, psl_gam, dem = NA)
}
\arguments{
\item{ukcp_psl}{a \code{spatialarray} of daily sea-level pressure (Pa)}

\item{psl_gam}{a \code{gam} object of correction coefficients to apply to
specific humdidity data as returned by \code{\link[=gamcorrect]{gamcorrect()}}}

\item{dem}{an optional digital elevation dataset covering the same extent and
of the same resolution as data in ukcp_psl. Can be supplied as a raster or matrix
(see details)}
}
\value{
a spatial array of hourly pressure in Pa
}
\description{
Calculate hourly corrected sea-level pressure from daily data
}
\details{
This function applies a correction to \code{ukcp_psl} using
parameters in \code{psl_gam} and then interpolates daily data to hourly.
If digital elevation data are supplied, sea-level pressure is elevation corrected.
}
\seealso{
\code{\link[=gamcorrect]{gamcorrect()}}
}
