% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{hourlysis}
\alias{hourlysis}
\title{Calculate hourly radiation from daily radiation}
\usage{
hourlysis(dailysis, modfit, startyear, rad_gam = NA, Trace = T)
}
\arguments{
\item{dailysis}{an object of class spatialarray with hourly radiation values}

\item{modfit}{a list of semivariagram model coefficients as returned by \code{\link[=radfit]{radfit()}}}

\item{Trace}{logical value indicating whether to plot progress}
}
\value{
an object of class spatialarray containing the following components:
\describe{
\item{arraydata}{a three-dimensional array of hourly radiation values}
\item{times}{\code{POSIXlt} object of times associated with \code{arraydata}}
\item{crs}{\code{crs} object of coordinate reference system associated with \code{arraydata}}
\item{extent}{\code{extent} object giving extent covered by \code{arraydata}}
\item{units}{units of \code{arraydata}}
\item{description}{character description of \code{arraydata}}
}
}
\description{
Calculate hourly radiation from daily radiation
}
\details{
Hourly radiation values are computed by deriving daily optical depths
from daily radiation values, interpolating these to hourly and back-calculating
radiation for each hour. To accunt for sub-daily variation in cloud patchiness
spatially and temporally autocorrelated cloud patchiness is simulated using
paramaters from \code{modfit}.
}
\examples{
# Takes a few seconds to run
modfit <- radfit(sis_2005)
# Takes a few minutes to run
ukcpsishourly <- hourlysis(sis_ukcpdaily, modfit, 2000)
}
\seealso{
\code{\link[=radfit]{radfit()}}
}
