% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{hourlytc}
\alias{hourlytc}
\title{Calculate hourly temperature from daily max and min}
\usage{
hourlytc(huss, pre, cfc, dni, dif, tmin, tmax, dem, dtr_gam, tc_gam)
}
\arguments{
\item{huss}{a \code{spatialarray} of hourly specific humidities (Kg / Kg) as returned by \code{\link[=hourlyhuss]{hourlyhuss()}}}

\item{pre}{a \code{spatialarray} of hourly pressures (Pa) as returned by \code{\link[=hourlypsl]{hourlypsl()}}}

\item{cfc}{a \code{spatialarray} of hourly percentage cloud cover as returned by \code{\link[=radsplit]{radsplit()}}}

\item{dni}{a \code{spatialarray} of hourly direct normal irradiances (W / m^2) as returned by \code{\link[=radsplit]{radsplit()}}}

\item{dif}{a \code{spatialarray} of hourly horizontal diffuse radiation (W / m^2) as returned by \code{\link[=radsplit]{radsplit()}}}

\item{tmin}{a \code{spatialarray} of daily minimum temperature (deg C)}

\item{tmax}{a \code{spatialarray} of daily maximum temperature (deg C)}

\item{dem}{a raster of eleveations of the same resolution and extent as \code{huss}}

\item{dtr_gam}{a \code{gam} object of correction coefficients to apply to
diurnal temperature ranges as returned by \code{\link[=gamcorrect]{gamcorrect()}}}

\item{tc_gam}{a \code{gam} object of correction coefficients to apply to
mean temperatures as returned by \code{\link[=gamcorrect]{gamcorrect()}}}
}
\value{
a \code{spatialarray} of hourly temperatures (deg C)
}
\description{
Calculate hourly temperature from daily max and min
}
\details{
This function uses the \code{\link[=hourlytemp]{hourlytemp()}} in the \code{microclima} package to compute
hourly temperatures. Percentage cloud cover is automatically converted to fractional
cloud cover, and units of radiation are automatically converted to MJ / m^2 / hr
}
\seealso{
\code{\link[=hourlyhuss]{hourlyhuss()}} \code{\link[=hourlypsl]{hourlypsl()}} \code{\link[=radsplit]{radsplit()}} \code{\link[=gamcorrect]{gamcorrect()}}  \code{\link[microclima:hourlytemp]{microclima::hourlytemp()}}
}
