% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{hourlywind}
\alias{hourlywind}
\title{Calculate hourly corrected wind speed and direction from daily data}
\usage{
hourlywind(ukcp_uw, ukcp_vw, ws_gam)
}
\arguments{
\item{ukcp_uw}{a \code{spatialarray} of u vector of daily wind speed (m / s)}

\item{ws_gam}{a \code{gam} object of correction coefficients to apply to
wind speed data as returned by \code{\link[=gamcorrect]{gamcorrect()}}}

\item{ukcp_uw}{a \code{spatialarray} of v vector of daily wind speed (m / s)}
}
\value{
a list of two spatialarrays
\describe{
\item{ws}{A \code{spatialarray} of hourly wind speed (m /s)}
\item{wd}{A \code{spatialarray} of hourly wind direction (degrees)}
}
}
\description{
Calculate hourly corrected wind speed and direction from daily data
}
\details{
This function interpolates u and v wind vectors to hourly and then
calculates wind speed and direction. It then applies a correction
using parameters in \code{ws_gam}.
}
\seealso{
\code{\link[=gamcorrect]{gamcorrect()}}
}
