% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hus_gam}
\alias{hus_gam}
\title{General Additive Model cofficients for correcting UKCP18 specific humidity data}
\format{An object of class \code{gam} (inherits from \code{glm}, \code{lm}) of length 46.}
\usage{
hus_gam
}
\description{
An object of class \code{gam} containing coefficients for correcting UKCP18 specific
humidity data, where specific humdidity in Kg / Kg has been multiplied by 100,000,
as returned by \code{\link[=gamcorrect]{gamcorrect()}}.
}
\keyword{datasets}
