% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthlysst}
\alias{monthlysst}
\title{Monthly sea-surface temperatures in Cornwall}
\format{An object of class \code{spatialarray} of length 6.}
\source{
\url{www.esrl.noaa.gov/}
}
\usage{
monthlysst
}
\description{
An object of class \code{spatialarray} of monthly sea-surface temperature anomolies in Cornwall in 2017
}
\details{
#' @format A list with the following components:
\describe{
\item{arraydata}{a three-dimensional array of monthlysea-surface temperatures }
\item{times}{\code{POSIXlt} object of times associated with \code{arraydata}}
\item{crs}{\code{crs} object of coordinate reference system associated with \code{arraydata}}
\item{extent}{\code{extent} object giving extent covered by \code{arraydata}}
\item{units}{units of \code{arraydata}}
\item{description}{character description of \code{arraydata}}
}
}
\keyword{datasets}
