% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msk}
\alias{msk}
\title{5 km raster of Cornwall}
\format{A raster with 23 rows and 34 columns:}
\usage{
msk
}
\description{
A 5 km resolution raster of Cornwall with land coded as 1 and sea as NA.
}
\keyword{datasets}
