% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{opticaldepth}
\alias{opticaldepth}
\title{calculate optical depth from radiation}
\usage{
opticaldepth(rad)
}
\arguments{
\item{rad}{\code{spatialarray} of total surface incoming radiation}
}
\value{
a \code{spatialarray} of optical depths
}
\description{
calculate optical depth from radiation
}
\examples{
od <- opticaldepth(sis_2005)
}
