% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pre_gam}
\alias{pre_gam}
\title{General Additive Model cofficients for correcting UKCP18 log (precipitation data)}
\format{An object of class \code{gam} (inherits from \code{glm}, \code{lm}) of length 46.}
\usage{
pre_gam
}
\description{
An object of class \code{gam} containing coefficients for correcting log of UKCP18 daily precipitation
data, as returned by \code{\link[=gamcorrect]{gamcorrect()}}.
}
\keyword{datasets}
