% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rad_gam}
\alias{rad_gam}
\title{General Additive Model cofficients for correcting UKCP18 radiation data}
\format{An object of class \code{gam} (inherits from \code{glm}, \code{lm}) of length 46.}
\usage{
rad_gam
}
\description{
An object of class \code{gam} containing coefficients for correcting UKCP18 radiation
data, as returned by \code{\link[=gamcorrect]{gamcorrect()}}.
}
\keyword{datasets}
