% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{radbyyear}
\alias{radbyyear}
\title{Calculate and save corrected radiation by year}
\usage{
radbyyear(ukcpsishourly, rad_gam, years, destfolder)
}
\arguments{
\item{ukcpsishourly}{a \code{spatialarray} of hourly uncorrected
surface incoming solar irradiance (W / m^2) as returned by \code{\link[=hourlysis]{hourlysis()}}}

\item{rad_gam}{a \code{gam} object of correction coefficients to apply to
radiation data as returned by \code{\link[=gamcorrect]{gamcorrect()}}}

\item{years}{years for which data are required. Must overlap with years
in \code{ukcpsishourly}}

\item{destfolder}{folder in which to save data (see details).If it does
not contain an absolute path, the directory name is relative to the
current working directory, \code{\link[=getwd]{getwd()}}. Tilde-expansion is performed where
supported.}
}
\description{
Calculate and save corrected radiation by year
}
\details{
This function applies a correction to \code{ukcpsishourly} using
parameters in \code{rad_gam}. Total surface incoming solar irradiance is then
aparitioned between direct and diffuse, and direct irradiance normal to
solar beam computed. A cloud cover coefficient is also computed from the
ratio of received to clearsky radiation. The resulting data are saved seperately
by year as \code{spatialarrays} in three subfolders of \code{destfolder}.
}
\examples{
# Takes a few seconds to run
modfit <- radfit(sis_2005)
# Takes a few minutes to run
ukcpsishourly <- hourlysis(sis_ukcpdaily, modfit, 2000)
yrs <- c(2001:2009)
radbyyear(ukcpsishourly, rad_gam, yrs, destfolder = "C:/radiation/")
}
\seealso{
\code{\link[=hourlysis]{hourlysis()}} \code{\link[=gamcorrect]{gamcorrect()}} \code{\link[microclima:difprop]{microclima::difprop()}}
}
