% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{radfit}
\alias{radfit}
\title{Calculate hourly radiation patchiness coefficients}
\usage{
radfit(sis)
}
\arguments{
\item{sis}{an object of class spatial array with hourly radiation values}
}
\value{
a list containing the following components:
\describe{
\item{m1}{A list of semivariagram model coefficients for the six-hourly deviations from daily values}
\item{m2}{A list of semivariagram model coefficients for the hourly deviations from six-hourly values}
}
}
\description{
Calculate hourly radiation patchiness coefficients
}
\details{
Using \code{\link[=hourlysis]{hourlysis()}}, hourly radiation values are computed by deriving
daily optical depths from daily radiation values, interpolating these to hourly
and back-calculating radiation for each hour. In so doing, cloud patchiness is
under-estimated, leading to potential errors in the calculation of direct and
diffuse radiation. The function \code{\link[=radfit]{radfit()}} derives model coefficients from real
hourly radiation data so that spatially and temporally autocorrelated cloud
patchiness can be simulated.
}
\examples{
modfit <- radfit(sis_2005)
modfit

}
\seealso{
\code{\link[=hourlysis]{hourlysis()}}
}
