% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{radsplit}
\alias{radsplit}
\title{Partition radiation into its direct and diffuse components}
\usage{
radsplit(ukcpsishourly)
}
\arguments{
\item{ukcpsishourly}{a \code{spatialarray} of hourly surface incoming solar irradiance
(W / m^2) as returned by \code{\link[=hourlysis]{hourlysis()}}}
}
\value{
a list of three spatialarrays
\describe{
\item{dni}{A \code{spatialarray} of direct radiation normal to the solar beam (W / m^2)}
\item{dif}{A \code{spatialarray} of horizontal diffuse radiation (W / m^2)}
\item{cfc}{A \code{spatialarray} of cloud cover derived as the ratio of actual to
clear-sky radiation (Percentage)}
}
}
\description{
Partition radiation into its direct and diffuse components
}
\examples{
# Takes a few seconds to run
modfit <- radfit(sis_2005)
# Takes a few minutes to run
ukcpsishourly <- hourlysis(sis_ukcpdaily, modfit, rad_gam)
radcom <- radsplit(ukcpsishourly)
attributes(radcom)
}
\seealso{
\code{\link[=hourlysis]{hourlysis()}} \code{\link[microclima:difprop]{microclima::difprop()}}
}
