% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{rain_adj}
\alias{rain_adj}
\title{Adjusts ukcp daily precipitation data to make it more consistent with observed data}
\usage{
rain_adj(ukcp_pre, rzero = 18.59483, pre_gam)
}
\arguments{
\item{ukcp_pre}{a \code{spatialarray} of UKCP18 daily rainfall data as returned by \code{\link[=cropresamplerain]{cropresamplerain()}}}

\item{rzero}{an optional single numeric value corresponding to the  ratio of observed to ukcp18
days with zero precipitation as returned by \code{\link[=rainzero_adj]{rainzero_adj()}}.}

\item{pre_gam}{a \code{gam} object of correction coefficients to apply to
precipitation data as returned by [gamcorrect()}
}
\value{
a \code{spatialarray} of daily rainfall in mm
}
\description{
Adjusts ukcp daily precipitation data to make it more consistent with observed data
}
\details{
This function applies a correction to \code{ukcp_pre} using
parameters in \code{pre_gam} and then interpolates daily data to hourly.
}
\seealso{
\code{\link[=gamcorrect]{gamcorrect()}}
}
