% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{rainapplyzero}
\alias{rainapplyzero}
\title{Adjusts UKCP18 precipitation data to give expected proportion of days with zero precipitation}
\usage{
rainapplyzero(ukcp, rzero = 18.59483)
}
\arguments{
\item{ukcp}{a \code{spatialarray} of UKCP18 daily rainfall data as returned by \code{\link[=cropresamplerain]{cropresamplerain()}}}

\item{rzero}{an optional single numeric value corresponding to the  ratio of observed to ukcp18
days with zero precipitation as returned by \code{\link[=rainzero_adj]{rainzero_adj()}}.}
}
\description{
Adjusts UKCP18 precipitation data to give expected proportion of days with zero precipitation
}
\details{
if no value for \code{rzero} is given, a default value derived for Cornwall, UK for
the period 2000-12-01 to 2010-11-30 is used. If \code{rzero} > 1 (likely to be almost always true),
then the requisite number of zeros is assigned to days with the lowest rainfall. If \code{rzero} < 1
then a dataset of precipitation values with the same mean and standard deviation of non-zero
precipitations values is generated, and the requisite number of days with zero rainfall are
assigned the lowest randomly generated precipitations values.
}
