% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{rainzero_adj}
\alias{rainzero_adj}
\title{Calculates ratio of observed to ukcp18 days with zero precipitation}
\usage{
rainzero_adj(observed_pre, ukcp_pre)
}
\arguments{
\item{observed_pre}{a \code{spatialarray} of observed daily precipitation values}

\item{ukcp_pre}{a \code{spatialarray} of ukcp daily precipitation values for the period
and extent corresponding to \code{observed}}
}
\value{
a single numeric value of the proportion of days with zero precipitation in
the observed dataset divided by the corresponding proportion of days in the UKCP18 data
}
\description{
Calculates ratio of observed to ukcp18 days with zero precipitation
}
