% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{savebyyear}
\alias{savebyyear}
\title{Saves data by year in specified directory}
\usage{
savebyyear(sa, yrs, dirout, filename)
}
\arguments{
\item{sa}{a \code{spatialarray} of data spanning the period in \code{yrs}}

\item{yrs}{a vector of years}

\item{dirout}{directory in which to save data. If it does not contain an absolute
path, the directory  name is relative to the current working directory, \code{\link[=getwd]{getwd()}}.
Tilde-expansion is performed where supported.}

\item{filename}{file name of data to be saved. Each file is saved a sa \code{spatialarray}
with the following name: \code{filenameyear.R}}
}
\description{
Saves data by year in specified directory
}
