% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sis_2005}
\alias{sis_2005}
\title{Hourly surface incoming shortwave radiation (observed).}
\format{A list with the following components:
\describe{
\item{arraydata}{a three-dimensional array of hourly radiation values}
\item{times}{\code{POSIXlt} object of times associated with \code{arraydata}}
\item{crs}{\code{crs} object of coordinate reference system associated with \code{arraydata}}
\item{extent}{\code{extent} object giving extent covered by \code{arraydata}}
\item{units}{units of \code{arraydata}}
\item{description}{character description of \code{arraydata}}
}}
\source{
\url{https://www.cmsaf.eu/}
}
\usage{
sis_2005
}
\description{
An object of class \code{spatialarray} containing hourly radiation values in Cornwall for
the whole of 2005
}
\keyword{datasets}
