% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{sst_fit}
\alias{sst_fit}
\title{Derive parameters relating land to sea temperatures}
\usage{
sst_fit(temps, Trace = TRUE)
}
\arguments{
\item{temps}{a data.frame of hourly temperatures with the following columns:
\describe{
\item{obs_time}{an object of class \code{POSIXlt} of observation times}
\item{land}{land temperatures (Deg C)}
\item{sst}{Sea-surface temperatures (Deg C)}
}}

\item{Trace}{an optional logical indicating whether to plot observed versus predicted
sea-surface temperatures}
}
\value{
a data frame with the following columns
\describe{
\item{x1}{parameter relating land-sea temperature difference to sea-surface temperature change}
\item{x2}{parameter relating mean sea-surfacetemperatures to sea-surface temperature change}
\item{x3}{a constant}
\item{k}{mean sea-surface temperature - mean land temperature}
}
}
\description{
Derive parameters relating land to sea temperatures
}
\examples{
sstparams <- sst_fit(temperature2017)
sstparams
}
\seealso{
\code{\link[=sst_predict]{sst_predict()}}
}
