% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{sst_predict}
\alias{sst_predict}
\title{Derive sea-surface temperatures from land temperatures}
\usage{
sst_predict(sstparams, htemp)
}
\arguments{
\item{sstparams}{of parameter values as returned by \code{\link[=sst_fit]{sst_fit()}}}

\item{htemps}{a vector of hourly land temperatues, normally for one year (deg C)}
}
\value{
a vector of hourly sea-surface temperatures
}
\description{
Derive sea-surface temperatures from land temperatures
}
\examples{
sstparams <- sst_fit(temperature2017)
htemp <- rep(temperature2017$land, each = 24)
sstp <- sst_predict(sstparams, htemp)
ssto <- rep(temperature2017$sst, each = 24)
plot(ssto, type = "l", ylim = c(5, 20), col = "red", ylab = "SST")
par(new = T)
plot(sstp, type = "l", ylim = c(5, 20), col = "darkgray", ylab = "")
}
\seealso{
\code{\link[=sst_fit]{sst_fit()}}
}
