% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{sst_spatial}
\alias{sst_spatial}
\title{Generates spatial array of sea-surface temperatures}
\usage{
sst_spatial(sstv, monthlysst, year)
}
\arguments{
\item{sstv}{vector of hourly sea surface temperatures over one year}

\item{monthlysst}{an object of class \code{spatialarray} of spatial anomolies in
sea-surface tempertaures in each month with the following components:
\describe{
\item{arraydata}{a three-dimensional array of monthlysea-surface temperatures }
\item{times}{\code{POSIXlt} object of times associated with \code{arraydata}}
\item{crs}{\code{crs} object of coordinate reference system associated with \code{arraydata}}
\item{extent}{\code{extent} object giving extent covered by \code{arraydata}}
\item{units}{units of \code{arraydata}}
\item{description}{character description of \code{arraydata}}
}}

\item{year}{year for which output data required. Used to derive correct number of
hourly values}
}
\value{
an object of class \code{spatialarray} of hourly sea-surface temperatures
}
\description{
Generates spatial array of sea-surface temperatures
}
