% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature2017}
\alias{temperature2017}
\title{A data frame of mean daily temperatures across cornwall}
\format{A dataframe with the following columns:
\describe{
\item{obs_time}{an object of class \code{POSIXlt} of observation times}
\item{land}{land temperatures (Deg C)}
\item{sst}{Sea-surface temperatures (Deg C)}
}}
\usage{
temperature2017
}
\description{
A data frame of mean daily temperatures across cornwall
}
\keyword{datasets}
