% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.timesel}
\alias{.timesel}
\title{reates time sequence matching 10 years of UKCP18 data, with missing days set as NA}
\usage{
.timesel(startyear)
}
\description{
reates time sequence matching 10 years of UKCP18 data, with missing days set as NA
}
