% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{hourlyhuss}
\alias{hourlyhuss}
\title{Calculate hourly corrected specific humidity from daily data}
\usage{
hourlyhuss(ukcp_huss, hus_gam, htemps = NA, hpre = NA)
}
\arguments{
\item{ukcp_huss}{a \code{spatialarray} of daily specific humidity (Kg / Kg)}

\item{hus_gam}{a \code{gam} object of correction coefficients to apply to
specific humdidity data as returned by \code{\link[=gamcorrect]{gamcorrect()}}}

\item{htemps}{optional spatial array of hourly temperatures (deg C - see details)}

\item{hpre}{optional spatial array  of hourly pressure (Pa - see details)}
}
\value{
a spatial array of hourly specific humidities in Kg / Kg
}
\description{
Calculate hourly corrected specific humidity from daily data
}
\details{
This function applies a correction to \code{ukcp_huss} using
parameters in \code{hus_gam} and then interpolates daily data to hourly.
If temperature data are supplied, hourly specific humdidities are conveted
to relative humidity, which is capped at 100 prior to converting back to
specific humdidity. Hourly pressure data are used in converting to relative
humidity. If not supplied, then pressure is assumed to be 101300 Pa.
}
\seealso{
\code{\link[=gamcorrect]{gamcorrect()}}
}
